fix 1006

+++ app/Http/Controllers/BaiVietController.php
@@ -4,14 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\createBaiVietRequest;
 use App\Models\BaiViet;
+use App\Models\ChiTietPhanQuyen;
+use App\Models\ChuyenMuc;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BaiVietController extends Controller
 {
     public function timKiem(Request $request)
     {
         $id_chuc_nang   = 53;
-
+        $user   =  Auth::guard('sanctum')->user();
+        $check  =   ChiTietPhanQuyen::where('id_quyen', $user->id_chuc_vu)
+                                    ->where('id_chuc_nang', $id_chuc_nang)
+                                    ->first();
+        if(!$check) {
+            return response()->json([
+                'status'    =>  false,
+                'message'   =>  'Bạn không đủ quyền truy cập chức năng này!',
+            ]);
+        }
+        
         $noi_dung   = '%' . $request->noi_dung_tim . '%';
 
         $data   = BaiViet::where('ten_bai_viet', 'like', $noi_dung)
@@ -37,15 +50,29 @@ class BaiVietController extends Controller
     public function getData()
     {
         $id_chuc_nang   = 48;
-
-        $data = BaiViet::all();
+        $user   =  Auth::guard('sanctum')->user();
+        $check  =   ChiTietPhanQuyen::where('id_quyen', $user->id_chuc_vu)
+                                    ->where('id_chuc_nang', $id_chuc_nang)
+                                    ->first();
+        if(!$check) {
+            return response()->json([
+                'status'    =>  false,
+                'message'   =>  'Bạn không đủ quyền truy cập chức năng này!',
+            ]);
+        }
+        
+        $data = BaiViet::join('chuyen_mucs', 'chuyen_mucs.id', 'bai_viets.id_chuyen_muc')
+                       ->select('bai_viets.*', 'chuyen_mucs.ten_chuyen_muc')
+                       ->get();
         return response()->json([
             'bai_viet'  =>  $data
         ]);
     }
-    public function getdataClient()
+    public function getdataClient($slug_chuyen_muc)
     {
+        $chuyenMuc = ChuyenMuc::where('slug_chuyen_muc', $slug_chuyen_muc)->first();
         $data = BaiViet::where('tinh_trang', 1)
+                        ->where('id_chuyen_muc', $chuyenMuc->id)
                         ->select('bai_viets.*')
                         ->get(); // get là ra 1 danh sách;
 
@@ -57,7 +84,17 @@ class BaiVietController extends Controller
     public function store(createBaiVietRequest $request)
     {
         $id_chuc_nang   = 49;
-
+        $user   =  Auth::guard('sanctum')->user();
+        $check  =   ChiTietPhanQuyen::where('id_quyen', $user->id_chuc_vu)
+                                    ->where('id_chuc_nang', $id_chuc_nang)
+                                    ->first();
+        if(!$check) {
+            return response()->json([
+                'status'    =>  false,
+                'message'   =>  'Bạn không đủ quyền truy cập chức năng này!',
+            ]);
+        }
+        
         $data   =   $request->all();
         BaiViet::create($data);
         return response()->json([
@@ -69,7 +106,17 @@ class BaiVietController extends Controller
     public function destroy($id)
     {
         $id_chuc_nang   = 50;
-
+        $user   =  Auth::guard('sanctum')->user();
+        $check  =   ChiTietPhanQuyen::where('id_quyen', $user->id_chuc_vu)
+                                    ->where('id_chuc_nang', $id_chuc_nang)
+                                    ->first();
+        if(!$check) {
+            return response()->json([
+                'status'    =>  false,
+                'message'   =>  'Bạn không đủ quyền truy cập chức năng này!',
+            ]);
+        }
+        
         BaiViet::find($id)->delete();
 
         return response()->json([
@@ -81,7 +128,17 @@ class BaiVietController extends Controller
     public function update(Request $request)
     {
         $id_chuc_nang   = 51;
-
+        $user   =  Auth::guard('sanctum')->user();
+        $check  =   ChiTietPhanQuyen::where('id_quyen', $user->id_chuc_vu)
+                                    ->where('id_chuc_nang', $id_chuc_nang)
+                                    ->first();
+        if(!$check) {
+            return response()->json([
+                'status'    =>  false,
+                'message'   =>  'Bạn không đủ quyền truy cập chức năng này!',
+            ]);
+        }
+        
         $data   = $request->all();
         BaiViet::find($request->id)->update($data);
 
@@ -94,7 +151,17 @@ class BaiVietController extends Controller
     public function doiTrangThai(Request $request)
     {
         $id_chuc_nang   = 52;
-
+        $user   =  Auth::guard('sanctum')->user();
+        $check  =   ChiTietPhanQuyen::where('id_quyen', $user->id_chuc_vu)
+                                    ->where('id_chuc_nang', $id_chuc_nang)
+                                    ->first();
+        if(!$check) {
+            return response()->json([
+                'status'    =>  false,
+                'message'   =>  'Bạn không đủ quyền truy cập chức năng này!',
+            ]);
+        }
+        
         $bai_viet = BaiViet::find($request->id);
         if($bai_viet) {
             if($bai_viet->tinh_trang == 1) {
